+++ tomcat/src/main/java/org/apache/coyote/connector/Connector.java
@@ -1,6 +1,6 @@
-package org.apache.catalina.connector;
+package org.apache.coyote.connector;
 
-import org.apache.coyote.http11.Http11Processor;
+import org.apache.coyote.processor.Http11Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
