+++ tomcat/src/main/java/org/apache/coyote/coockie/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.coyote.coockie;
 
 import java.util.HashMap;
 import java.util.Map;
