+++ tomcat/src/main/java/org/apache/coyote/http/ResourceType.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.coyote.http;
 
 public enum ResourceType {
     NON_STATIC, STATIC
