+++ tomcat/src/main/java/org/apache/coyote/http/StatusCode.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.coyote.http;
 
 public enum StatusCode {
     _200(200, "OK"),
