+++ tomcat/src/main/java/org/apache/coyote/processor/Http11Processor.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.processor;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.exception.UncheckedServletException;
@@ -6,15 +6,15 @@ import com.techcourse.model.User;
 import java.io.IOException;
 import java.net.Socket;
 import java.util.Optional;
-import org.apache.catalina.HeaderName;
-import org.apache.catalina.HttpMethod;
-import org.apache.catalina.HttpRequest;
-import org.apache.catalina.HttpResponse;
-import org.apache.catalina.RequestBody;
-import org.apache.catalina.ResourceType;
-import org.apache.catalina.Session;
-import org.apache.catalina.StatusCode;
-import org.apache.catalina.manager.SessionManager;
+import org.apache.coyote.http.HeaderName;
+import org.apache.coyote.http.HttpMethod;
+import org.apache.coyote.http.ResourceType;
+import org.apache.coyote.http.StatusCode;
+import org.apache.coyote.request.HttpRequest;
+import org.apache.coyote.response.HttpResponse;
+import org.apache.coyote.request.RequestBody;
+import org.apache.coyote.session.Session;
+import org.apache.coyote.manager.SessionManager;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
