+++ tomcat/src/main/java/org/apache/coyote/request/HttpRequest.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.coyote.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -6,6 +6,10 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.HashMap;
 import java.util.Map;
+import org.apache.coyote.http.ContentType;
+import org.apache.coyote.http.HeaderName;
+import org.apache.coyote.coockie.HttpCookie;
+import org.apache.coyote.http.HttpMethod;
 
 public class HttpRequest {
